move ReadFile and Stat commands from git package to gitserver client
