+++ cmd/frontend/graphqlbackend/repository_comparison.go
@@ -20,7 +20,6 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/api"
 	"github.com/sourcegraph/sourcegraph/internal/database"
 	"github.com/sourcegraph/sourcegraph/internal/gitserver"
-	"github.com/sourcegraph/sourcegraph/internal/vcs/git"
 	"github.com/sourcegraph/sourcegraph/lib/errors"
 )
 
@@ -78,7 +77,7 @@ func NewRepositoryComparison(ctx context.Context, db database.DB, r *RepositoryR
 
 	client := gitserver.NewClient(db)
 	getCommit := func(ctx context.Context, repo api.RepoName, revspec string) (*GitCommitResolver, error) {
-		if revspec == git.DevNullSHA {
+		if revspec == gitserver.DevNullSHA {
 			return nil, nil
 		}
