+++ internal/gitserver/commands_test.go
@@ -4,8 +4,11 @@ import (
 	"context"
 	"fmt"
 	"io"
+	"io/fs"
 	"os"
+	"path/filepath"
 	"reflect"
+	"runtime"
 	"sort"
 	"strings"
 	"testing"
@@ -976,5 +979,161 @@ func TestMerger_MergeBase(t *testing.T) {
 			t.Errorf("%s: MergeBase(%s, %s): got %q, want %q", label, a, b, mb, want)
 			continue
 		}
+	}
+}
+
+func TestRepository_FileSystem_Symlinks(t *testing.T) {
+	ClientMocks.LocalGitserver = true
+	defer ResetClientMocks()
+
+	db := database.NewMockDB()
+	gitCommands := []string{
+		"touch file1",
+		"mkdir dir1",
+		"ln -s file1 link1",
+		"ln -s ../file1 dir1/link2",
+		"touch --date=2006-01-02T15:04:05Z file1 link1 dir1/link2 || touch -t " + Times[0] + " file1 link1 dir1/link2",
+		"git add link1 file1 dir1/link2",
+		"GIT_COMMITTER_NAME=a GIT_COMMITTER_EMAIL=a@a.com GIT_COMMITTER_DATE=2006-01-02T15:04:05Z git commit -m commit1 --author='a <a@a.com>' --date 2006-01-02T15:04:05Z",
+	}
+
+	// map of path to size of content
+	symlinks := map[string]int64{
+		"link1":      5, // file1
+		"dir1/link2": 8, // ../file1
+	}
+
+	dir := InitGitRepository(t, gitCommands...)
+	repo := api.RepoName(filepath.Base(dir))
+
+	client := NewClient(db)
+
+	commitID := api.CommitID(ComputeCommitHash(dir, true))
+
+	ctx := context.Background()
+
+	// file1 should be a file.
+	file1Info, err := client.Stat(ctx, authz.DefaultSubRepoPermsChecker, repo, commitID, "file1")
+	if err != nil {
+		t.Fatalf("fs.Stat(file1): %s", err)
+	}
+	if !file1Info.Mode().IsRegular() {
+		t.Errorf("file1 Stat !IsRegular (mode: %o)", file1Info.Mode())
+	}
+
+	checkSymlinkFileInfo := func(name string, link fs.FileInfo) {
+		t.Helper()
+		if link.Mode()&os.ModeSymlink == 0 {
+			t.Errorf("link mode is not symlink (mode: %o)", link.Mode())
+		}
+		if link.Name() != name {
+			t.Errorf("got link.Name() == %q, want %q", link.Name(), name)
+		}
+	}
+
+	// Check symlinks are links
+	for symlink := range symlinks {
+		fi, err := client.LStat(ctx, authz.DefaultSubRepoPermsChecker, repo, commitID, symlink)
+		if err != nil {
+			t.Fatalf("fs.lStat(%s): %s", symlink, err)
+		}
+		if runtime.GOOS != "windows" {
+			// TODO(alexsaveliev) make it work on Windows too
+			checkSymlinkFileInfo(symlink, fi)
+		}
+	}
+
+	// Also check the FileInfo returned by ReadDir to ensure it's
+	// consistent with the FileInfo returned by lStat.
+	entries, err := client.ReadDir(ctx, db, authz.DefaultSubRepoPermsChecker, repo, commitID, ".", false)
+	if err != nil {
+		t.Fatalf("fs.ReadDir(.): %s", err)
+	}
+	found := false
+	for _, entry := range entries {
+		if entry.Name() == "link1" {
+			found = true
+			if runtime.GOOS != "windows" {
+				checkSymlinkFileInfo("link1", entry)
+			}
+		}
+	}
+	if !found {
+		t.Fatal("readdir did not return link1")
+	}
+
+	for symlink, size := range symlinks {
+		fi, err := client.Stat(ctx, authz.DefaultSubRepoPermsChecker, repo, commitID, symlink)
+		if err != nil {
+			t.Fatalf("fs.Stat(%s): %s", symlink, err)
+		}
+		if fi.Mode()&fs.ModeSymlink == 0 {
+			t.Errorf("%s Stat is not a symlink (mode: %o)", symlink, fi.Mode())
+		}
+		if fi.Name() != symlink {
+			t.Errorf("got Name %q, want %q", fi.Name(), symlink)
+		}
+		if fi.Size() != size {
+			t.Errorf("got %s Size %d, want %d", symlink, fi.Size(), size)
+		}
+	}
+}
+
+func TestStat(t *testing.T) {
+	ClientMocks.LocalGitserver = true
+	defer ResetClientMocks()
+
+	db := database.NewMockDB()
+	gitCommands := []string{
+		"mkdir dir1",
+		"touch dir1/file1",
+		"git add dir1/file1",
+		"GIT_COMMITTER_NAME=a GIT_COMMITTER_EMAIL=a@a.com GIT_COMMITTER_DATE=2006-01-02T15:04:05Z git commit -m commit1 --author='a <a@a.com>' --date 2006-01-02T15:04:05Z",
+	}
+
+	dir := InitGitRepository(t, gitCommands...)
+	repo := api.RepoName(filepath.Base(dir))
+	client := NewClient(db)
+
+	commitID := api.CommitID(ComputeCommitHash(dir, true))
+
+	ctx := context.Background()
+
+	checker := authz.NewMockSubRepoPermissionChecker()
+	// Start disabled
+	checker.EnabledFunc.SetDefaultHook(func() bool {
+		return false
+	})
+
+	fileInfo, err := client.Stat(ctx, checker, repo, commitID, "dir1/file1")
+	if err != nil {
+		t.Fatal(err)
+	}
+	want := "dir1/file1"
+	if diff := cmp.Diff(want, fileInfo.Name()); diff != "" {
+		t.Fatal(diff)
+	}
+
+	// With filtering
+	checker.EnabledFunc.SetDefaultHook(func() bool {
+		return true
+	})
+	checker.PermissionsFunc.SetDefaultHook(func(ctx context.Context, i int32, content authz.RepoContent) (authz.Perms, error) {
+		if strings.HasPrefix(content.Path, "dir2") {
+			return authz.Read, nil
+		}
+		return authz.None, nil
+	})
+	ctx = actor.WithActor(ctx, &actor.Actor{
+		UID: 1,
+	})
+
+	_, err = client.Stat(ctx, checker, repo, commitID, "dir1/file1")
+	if err == nil {
+		t.Fatal(err)
+	}
+	want = "ls-tree dir1/file1: file does not exist"
+	if diff := cmp.Diff(want, err.Error()); diff != "" {
+		t.Fatal(diff)
 	}
 }
