+++ internal/vcs/git/tree_test.go
@@ -7,12 +7,8 @@ import (
 	"os"
 	"path/filepath"
 	"reflect"
-	"runtime"
 	"sort"
-	"strings"
 	"testing"
-
-	"github.com/google/go-cmp/cmp"
 
 	"github.com/sourcegraph/sourcegraph/internal/actor"
 	"github.com/sourcegraph/sourcegraph/internal/api"
@@ -76,107 +72,6 @@ func TestReadDir_SubRepoFiltering(t *testing.T) {
 	}
 }
 
-func TestRepository_FileSystem_Symlinks(t *testing.T) {
-	t.Parallel()
-
-	db := database.NewMockDB()
-	gitCommands := []string{
-		"touch file1",
-		"mkdir dir1",
-		"ln -s file1 link1",
-		"ln -s ../file1 dir1/link2",
-		"touch --date=2006-01-02T15:04:05Z file1 link1 dir1/link2 || touch -t " + Times[0] + " file1 link1 dir1/link2",
-		"git add link1 file1 dir1/link2",
-		"GIT_COMMITTER_NAME=a GIT_COMMITTER_EMAIL=a@a.com GIT_COMMITTER_DATE=2006-01-02T15:04:05Z git commit -m commit1 --author='a <a@a.com>' --date 2006-01-02T15:04:05Z",
-	}
-
-	// map of path to size of content
-	symlinks := map[string]int64{
-		"link1":      5, // file1
-		"dir1/link2": 8, // ../file1
-	}
-
-	dir := InitGitRepository(t, gitCommands...)
-	repo := api.RepoName(filepath.Base(dir))
-
-	client := gitserver.NewClient(db)
-	if resp, err := client.RequestRepoUpdate(context.Background(), repo, 0); err != nil {
-		t.Fatal(err)
-	} else if resp.Error != "" {
-		t.Fatal(resp.Error)
-	}
-
-	commitID := api.CommitID(ComputeCommitHash(dir, true))
-
-	ctx := context.Background()
-
-	// file1 should be a file.
-	file1Info, err := Stat(ctx, db, authz.DefaultSubRepoPermsChecker, repo, commitID, "file1")
-	if err != nil {
-		t.Fatalf("fs.Stat(file1): %s", err)
-	}
-	if !file1Info.Mode().IsRegular() {
-		t.Errorf("file1 Stat !IsRegular (mode: %o)", file1Info.Mode())
-	}
-
-	checkSymlinkFileInfo := func(name string, link fs.FileInfo) {
-		t.Helper()
-		if link.Mode()&os.ModeSymlink == 0 {
-			t.Errorf("link mode is not symlink (mode: %o)", link.Mode())
-		}
-		if link.Name() != name {
-			t.Errorf("got link.Name() == %q, want %q", link.Name(), name)
-		}
-	}
-
-	// Check symlinks are links
-	for symlink := range symlinks {
-		fi, err := gitserver.NewClient(db).LStat(ctx, authz.DefaultSubRepoPermsChecker, repo, commitID, symlink)
-		if err != nil {
-			t.Fatalf("fs.lStat(%s): %s", symlink, err)
-		}
-		if runtime.GOOS != "windows" {
-			// TODO(alexsaveliev) make it work on Windows too
-			checkSymlinkFileInfo(symlink, fi)
-		}
-	}
-
-	// Also check the FileInfo returned by ReadDir to ensure it's
-	// consistent with the FileInfo returned by lStat.
-	entries, err := client.ReadDir(ctx, db, authz.DefaultSubRepoPermsChecker, repo, commitID, ".", false)
-	if err != nil {
-		t.Fatalf("fs.ReadDir(.): %s", err)
-	}
-	found := false
-	for _, entry := range entries {
-		if entry.Name() == "link1" {
-			found = true
-			if runtime.GOOS != "windows" {
-				checkSymlinkFileInfo("link1", entry)
-			}
-		}
-	}
-	if !found {
-		t.Fatal("readdir did not return link1")
-	}
-
-	for symlink, size := range symlinks {
-		fi, err := Stat(ctx, db, authz.DefaultSubRepoPermsChecker, repo, commitID, symlink)
-		if err != nil {
-			t.Fatalf("fs.Stat(%s): %s", symlink, err)
-		}
-		if fi.Mode()&fs.ModeSymlink == 0 {
-			t.Errorf("%s Stat is not a symlink (mode: %o)", symlink, fi.Mode())
-		}
-		if fi.Name() != symlink {
-			t.Errorf("got Name %q, want %q", fi.Name(), symlink)
-		}
-		if fi.Size() != size {
-			t.Errorf("got %s Size %d, want %d", symlink, fi.Size(), size)
-		}
-	}
-}
-
 func TestRepository_FileSystem(t *testing.T) {
 	t.Parallel()
 	ctx := context.Background()
@@ -216,15 +111,16 @@ func TestRepository_FileSystem(t *testing.T) {
 		},
 	}
 
+	client := gitserver.NewClient(db)
 	for label, test := range tests {
 		// notafile should not exist.
-		if _, err := Stat(ctx, db, authz.DefaultSubRepoPermsChecker, test.repo, test.first, "notafile"); !os.IsNotExist(err) {
+		if _, err := client.Stat(ctx, authz.DefaultSubRepoPermsChecker, test.repo, test.first, "notafile"); !os.IsNotExist(err) {
 			t.Errorf("%s: fs1.Stat(notafile): got err %v, want os.IsNotExist", label, err)
 			continue
 		}
 
 		// dir1 should exist and be a dir.
-		dir1Info, err := Stat(ctx, db, authz.DefaultSubRepoPermsChecker, test.repo, test.first, "dir1")
+		dir1Info, err := client.Stat(ctx, authz.DefaultSubRepoPermsChecker, test.repo, test.first, "dir1")
 		if err != nil {
 			t.Errorf("%s: fs1.Stat(dir1): %s", label, err)
 			continue
@@ -272,7 +168,7 @@ func TestRepository_FileSystem(t *testing.T) {
 		}
 
 		// dir1/file1 should exist, contain "infile1", have the right mtime, and be a file.
-		file1Data, err := ReadFile(ctx, db, test.repo, test.first, "dir1/file1", nil)
+		file1Data, err := client.ReadFile(ctx, test.repo, test.first, "dir1/file1", nil)
 		if err != nil {
 			t.Errorf("%s: fs1.ReadFile(dir1/file1): %s", label, err)
 			continue
@@ -280,7 +176,7 @@ func TestRepository_FileSystem(t *testing.T) {
 		if !bytes.Equal(file1Data, []byte("infile1")) {
 			t.Errorf("%s: got file1Data == %q, want %q", label, string(file1Data), "infile1")
 		}
-		file1Info, err = Stat(ctx, db, authz.DefaultSubRepoPermsChecker, test.repo, test.first, "dir1/file1")
+		file1Info, err = client.Stat(ctx, authz.DefaultSubRepoPermsChecker, test.repo, test.first, "dir1/file1")
 		if err != nil {
 			t.Errorf("%s: fs1.Stat(dir1/file1): %s", label, err)
 			continue
@@ -296,30 +192,30 @@ func TestRepository_FileSystem(t *testing.T) {
 		}
 
 		// file 2 shouldn't exist in the 1st commit.
-		_, err = ReadFile(ctx, db, test.repo, test.first, "file 2", nil)
+		_, err = client.ReadFile(ctx, test.repo, test.first, "file 2", nil)
 		if !os.IsNotExist(err) {
 			t.Errorf("%s: fs1.Open(file 2): got err %v, want os.IsNotExist (file 2 should not exist in this commit)", label, err)
 		}
 
 		// file 2 should exist in the 2nd commit.
-		_, err = ReadFile(ctx, db, test.repo, test.second, "file 2", nil)
+		_, err = client.ReadFile(ctx, test.repo, test.second, "file 2", nil)
 		if err != nil {
 			t.Errorf("%s: fs2.Open(file 2): %s", label, err)
 			continue
 		}
 
 		// file1 should also exist in the 2nd commit.
-		if _, err := Stat(ctx, db, authz.DefaultSubRepoPermsChecker, test.repo, test.second, "dir1/file1"); err != nil {
+		if _, err := client.Stat(ctx, authz.DefaultSubRepoPermsChecker, test.repo, test.second, "dir1/file1"); err != nil {
 			t.Errorf("%s: fs2.Stat(dir1/file1): %s", label, err)
 			continue
 		}
-		if _, err := ReadFile(ctx, db, test.repo, test.second, "dir1/file1", nil); err != nil {
+		if _, err := client.ReadFile(ctx, test.repo, test.second, "dir1/file1", nil); err != nil {
 			t.Errorf("%s: fs2.Open(dir1/file1): %s", label, err)
 			continue
 		}
 
 		// root should exist (via Stat).
-		root, err := Stat(ctx, db, authz.DefaultSubRepoPermsChecker, test.repo, test.second, ".")
+		root, err := client.Stat(ctx, authz.DefaultSubRepoPermsChecker, test.repo, test.second, ".")
 		if err != nil {
 			t.Errorf("%s: fs2.Stat(.): %s", label, err)
 			continue
@@ -432,7 +328,7 @@ func TestRepository_FileSystem_quoteChars(t *testing.T) {
 		}
 
 		for _, name := range wantNames {
-			stat, err := Stat(ctx, db, authz.DefaultSubRepoPermsChecker, test.repo, commitID, name)
+			stat, err := client.Stat(ctx, authz.DefaultSubRepoPermsChecker, test.repo, commitID, name)
 			if err != nil {
 				t.Errorf("%s: Stat(%q): %s", label, name, err)
 				continue
@@ -505,7 +401,7 @@ func TestRepository_FileSystem_gitSubmodules(t *testing.T) {
 
 		// Check the submodule fs.FileInfo both when it's returned by
 		// Stat and when it's returned in a list by ReadDir.
-		submod, err := Stat(ctx, db, authz.DefaultSubRepoPermsChecker, test.repo, commitID, "submod")
+		submod, err := client.Stat(ctx, authz.DefaultSubRepoPermsChecker, test.repo, commitID, "submod")
 		if err != nil {
 			t.Errorf("%s: fs.Stat(submod): %s", label, err)
 			continue
@@ -519,73 +415,10 @@ func TestRepository_FileSystem_gitSubmodules(t *testing.T) {
 		// .gitmodules file is entries[0]
 		checkSubmoduleFileInfo(label+" (ReadDir)", entries[1])
 
-		_, err = ReadFile(ctx, db, test.repo, commitID, "submod", nil)
+		_, err = client.ReadFile(ctx, test.repo, commitID, "submod", nil)
 		if err != nil {
 			t.Errorf("%s: fs.Open(submod): %s", label, err)
 			continue
 		}
-	}
-}
-
-func TestStat(t *testing.T) {
-	t.Parallel()
-
-	db := database.NewMockDB()
-	gitCommands := []string{
-		"mkdir dir1",
-		"touch dir1/file1",
-		"git add dir1/file1",
-		"GIT_COMMITTER_NAME=a GIT_COMMITTER_EMAIL=a@a.com GIT_COMMITTER_DATE=2006-01-02T15:04:05Z git commit -m commit1 --author='a <a@a.com>' --date 2006-01-02T15:04:05Z",
-	}
-
-	dir := InitGitRepository(t, gitCommands...)
-	repo := api.RepoName(filepath.Base(dir))
-
-	if resp, err := gitserver.NewClient(db).RequestRepoUpdate(context.Background(), repo, 0); err != nil {
-		t.Fatal(err)
-	} else if resp.Error != "" {
-		t.Fatal(resp.Error)
-	}
-
-	commitID := api.CommitID(ComputeCommitHash(dir, true))
-
-	ctx := context.Background()
-
-	checker := authz.NewMockSubRepoPermissionChecker()
-	// Start disabled
-	checker.EnabledFunc.SetDefaultHook(func() bool {
-		return false
-	})
-
-	fileInfo, err := Stat(ctx, db, checker, repo, commitID, "dir1/file1")
-	if err != nil {
-		t.Fatal(err)
-	}
-	want := "dir1/file1"
-	if diff := cmp.Diff(want, fileInfo.Name()); diff != "" {
-		t.Fatal(diff)
-	}
-
-	// With filtering
-	checker.EnabledFunc.SetDefaultHook(func() bool {
-		return true
-	})
-	checker.PermissionsFunc.SetDefaultHook(func(ctx context.Context, i int32, content authz.RepoContent) (authz.Perms, error) {
-		if strings.HasPrefix(content.Path, "dir2") {
-			return authz.Read, nil
-		}
-		return authz.None, nil
-	})
-	ctx = actor.WithActor(ctx, &actor.Actor{
-		UID: 1,
-	})
-
-	_, err = Stat(ctx, db, checker, repo, commitID, "dir1/file1")
-	if err == nil {
-		t.Fatal(err)
-	}
-	want = "ls-tree dir1/file1: file does not exist"
-	if diff := cmp.Diff(want, err.Error()); diff != "" {
-		t.Fatal(diff)
 	}
 }
